+++ gems.rb
@@ -15,6 +15,5 @@ end
 
 group :maintenance, optional: true do
 	gem "agent-context"
-	gem "bake-releases"
 	gem "utopia-project"
 end