+++ releases.md
@@ -1,6 +1,6 @@
 # Releases
 
-## Unreleased
+## v0.1.0
 
   - Include the version's release notes in GitHub releases using `bake-releases`.
   - Update generated release files in the working tree with `gem:github:setup:update`.